move console log filter to separate (serializable) class
